Clean up Grid structure and extract GridSample layout
